+++ src/main/kotlin/com/smody/book/security/authentication/OAuthLoginSuccessHandler.kt
@@ -1,9 +1,8 @@
-package com.smody.book.security
+package com.smody.book.security.authentication
 
 import com.fasterxml.jackson.databind.ObjectMapper
-import com.nimbusds.common.contenttype.ContentType
-import com.smody.book.member.domain.MemberRepository
-import com.smody.book.security.jwt.JwtTokenProvider
+import com.smody.book.security.OAuthPrincipal
+import com.smody.book.security.dto.TokenResponse
 import jakarta.servlet.http.HttpServletRequest
 import jakarta.servlet.http.HttpServletResponse
 import org.springframework.security.core.Authentication
@@ -12,8 +11,6 @@ import org.springframework.stereotype.Component
 
 @Component
 class OAuthLoginSuccessHandler(
-    private val memberRepository: MemberRepository,
-    private val jwtTokenProvider: JwtTokenProvider,
     private val objectMapper: ObjectMapper
 ) : AuthenticationSuccessHandler {
     override fun onAuthenticationSuccess(
