+++ src/main/kotlin/com/smody/book/security/dto/TokenResponse.kt
@@ -1,4 +1,6 @@
-package com.smody.book.security
+package com.smody.book.security.dto
+
+import com.smody.book.security.OAuthPrincipal
 
 data class TokenResponse(val memberId: Long, val accessToken: String) {
     constructor(oAuthPrincipal: OAuthPrincipal) : this(oAuthPrincipal.memberId, oAuthPrincipal.accessToken)
